began fixing more styling and reorganized dungeon components
