+++ src/itemTracker/items/dungeons/bk/fsBossKey.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import BK from '../../assets/dungeons/SS_Mysterious_Crystals_Icon.png'
-import No_BK from '../../assets/dungeons/fs_noBossKey.png'
+import BK from '../../../../assets/dungeons/SS_Mysterious_Crystals_Icon.png'
+import No_BK from '../../../../assets/dungeons/fs_noBossKey.png'
 
 export default class FS_BossKey extends React.Component {
     constructor(props) {
