+++ src/itemTracker/items/dungeons/bk/lmfBossKey.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import BK from '../../assets/dungeons/SS_Ancient_Circuit_Icon.png'
-import No_BK from '../../assets/dungeons/lmf_noBossKey.png'
+import BK from '../../../../assets/dungeons/SS_Ancient_Circuit_Icon.png'
+import No_BK from '../../../../assets/dungeons/lmf_noBossKey.png'
 
 export default class LMF_BossKey extends React.Component {
     constructor(props) {
