+++ src/itemTracker/items/dungeons/bk/sshBossKey.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import BK from '../../assets/dungeons/SS_Squid_Carving_Icon.png'
-import No_BK from '../../assets/dungeons/ssh_noBossKey.png'
+import BK from '../../../../assets/dungeons/SS_Squid_Carving_Icon.png'
+import No_BK from '../../../../assets/dungeons/ssh_noBossKey.png'
 
 export default class SSH_BossKey extends React.Component {
     constructor(props) {
