+++ src/itemTracker/items/dungeons/bk/stBossKey.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import BK from '../../assets/dungeons/SS_Golden_Carving_Icon.png'
-import No_BK from '../../assets/dungeons/st_noBossKey.png'
+import BK from '../../../../assets/dungeons/SS_Golden_Carving_Icon.png'
+import No_BK from '../../../../assets/dungeons/st_noBossKey.png'
 
 export default class ST_BossKey extends React.Component {
     constructor(props) {
