+++ src/itemTracker/items/dungeons/small/lmfSmall.js
@@ -1,9 +1,8 @@
 import React from 'react';
-import noKey from '../../assets/dungeons/noSmallKey.png'
-import smallKey_1 from '../../assets/dungeons/SS_Small_Key_Icon.png'
-import smallKey_2 from '../../assets/dungeons/2_smallKey.png'
+import noKey from '../../../../assets/dungeons/noSmallKey.png'
+import smallKey_1 from '../../../../assets/dungeons/SS_Small_Key_Icon.png'
 
-export default class SSH_Small extends React.Component {
+export default class LMF_Small extends React.Component {
     constructor(props) {
         super(props);
         this.handleClick = this.handleClick.bind(this);
@@ -14,16 +13,12 @@ export default class SSH_Small extends React.Component {
         let parent = this.props.parent
         switch(current) {
             case 0:
-                return <div id={"SSH-small-key"}>
+                return <div id={"LMF-small-key"}>
                     <img src={noKey} onClick={this.handleClick} alt={"No Small Key"}/>
                 </div>
             case 1:
-                return <div id={"SSH-small-key"}>
+                return <div id={"LMF-small-key"}>
                     <img src={smallKey_1} onClick={this.handleClick} alt={"Small Key 1"}/>
-                </div>
-            case 2:
-                return <div id={"SSH-small-key"}>
-                    <img src={smallKey_2} onClick={this.handleClick} alt={"Small Key 2"}/>
                 </div>
             default:
                 return null
@@ -31,6 +26,6 @@ export default class SSH_Small extends React.Component {
     }
 
     handleClick() {
-        this.props.onChange("sshSmall")
+        this.props.onChange("lmfSmall")
     }
 }
